move rest of the "mail setting" functionality from project-base to the framework package
